Move input components to common/input to free up elements directory for campaign-elements related stuff
