+++ src/js/client/components/common/input/checkbox.js
