+++ src/js/client/components/common/input/inputrange.js
