+++ src/js/client/components/preferences/accounts/AccountEditPage.js
@@ -1,7 +1,7 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
 import { Link } from 'react-router';
-import Input from '../../elements/input';
-import InputText from '../../elements/inputtext';
+import Input from '../../common/input/input';
+import InputText from '../../common/input/inputtext';
 import userService from '../../../services/UserService';
