+++ src/js/client/components/preferences/influencers/audience.js
@@ -1,6 +1,6 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
-import InputText from '../../elements/inputtext';
-import InputSelect from '../../elements/inputselect';
+import InputText from '../../common/input/inputtext';
+import InputSelect from '../../common/input/inputselect';
 
 const Audience = (props) => {
     if (!props.audience) {
