+++ src/js/client/components/projects/common/ProjectCheckpoint.js
@@ -1,5 +1,5 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
-import InputText from '../../elements/inputtext';
+import InputText from '../../common/input/inputtext';
 import moment from 'moment';
 
 const ProjectCheckpoint = (props) => {
