+++ src/js/client/components/search/components/personal.js
@@ -1,6 +1,6 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
-import InputText from '../../elements/inputtext';
-import InputSelect from '../../elements/inputselect';
+import InputText from '../../common/input/inputtext';
+import InputSelect from '../../common/input/inputselect';
 
 const Personal = (props) => {
     if (!props.personal) {
