+++ src/js/client/services/HistoryService.js
@@ -8,4 +8,4 @@ export default {
             })
             .catch(() => {});
     }
-}
+};
